Replace method division. Set signature to double and return one of operande by casting it to double type

+++ src/Calculs.java
@@ -22,9 +22,9 @@ public class Calculs
                 return operande1 + operande2;
         }
         
-        public int diviser()
+        public double diviser()
         {
-                return operande1 / operande2;
+                return (double)operande1 / operande2;
         }
         
         public int soustraire()
